extract almost all stuff to the platform-independent module
