+++ rust/gui-client/src-tauri/src/client/ipc.rs
@@ -1,16 +1,16 @@
 use crate::client::gui::{ControllerRequest, CtlrTx};
 use anyhow::{Context as _, Result};
 use arc_swap::ArcSwap;
-use connlib_client_shared::callbacks::ResourceDescription;
-use firezone_headless_client::IpcClientMsg;
-
+use connlib_client_shared::{callbacks::ResourceDescription, Callbacks};
+use firezone_headless_client::{IpcClientMsg, IpcServerMsg};
+use futures::{SinkExt, StreamExt};
+use secrecy::{ExposeSecret, SecretString};
 use std::{
     net::{IpAddr, Ipv4Addr, Ipv6Addr},
     sync::Arc,
 };
 use tokio::sync::Notify;
-
-pub(crate) use platform::Client;
+use tokio_util::codec::{FramedRead, FramedWrite, LengthDelimitedCodec};
 
 #[cfg(target_os = "linux")]
 #[path = "ipc/linux.rs"]
@@ -58,7 +58,75 @@ impl connlib_client_shared::Callbacks for CallbackHandler {
     }
 }
 
+pub(crate) struct Client {
+    task: tokio::task::JoinHandle<Result<()>>,
+    // Needed temporarily to avoid a big refactor. We can remove this in the future.
+    tx: FramedWrite<tokio::io::WriteHalf<platform::IpcStream>, LengthDelimitedCodec>,
+}
+
 impl Client {
+    pub(crate) async fn disconnect(mut self) -> Result<()> {
+        self.send_msg(&IpcClientMsg::Disconnect)
+            .await
+            .context("Couldn't send Disconnect")?;
+        self.tx.close().await?;
+        self.task.abort();
+        Ok(())
+    }
+
+    pub(crate) async fn send_msg(&mut self, msg: &IpcClientMsg) -> Result<()> {
+        self.tx
+            .send(
+                serde_json::to_string(msg)
+                    .context("Couldn't encode IPC message as JSON")?
+                    .into(),
+            )
+            .await
+            .context("Couldn't send IPC message")?;
+        Ok(())
+    }
+
+    pub(crate) async fn connect(
+        api_url: &str,
+        token: SecretString,
+        callback_handler: CallbackHandler,
+        tokio_handle: tokio::runtime::Handle,
+    ) -> Result<Self> {
+        tracing::info!(pid = std::process::id(), "Connecting to IPC service...");
+        let stream = platform::connect_to_service().await?;
+        let (rx, tx) = tokio::io::split(stream);
+        // Receives messages from the IPC service
+        let mut rx = FramedRead::new(rx, LengthDelimitedCodec::new());
+        let tx = FramedWrite::new(tx, LengthDelimitedCodec::new());
+
+        let task = tokio_handle.spawn(async move {
+            while let Some(msg) = rx.next().await.transpose()? {
+                match serde_json::from_slice::<IpcServerMsg>(&msg)? {
+                    IpcServerMsg::Ok => {}
+                    IpcServerMsg::OnDisconnect => callback_handler.on_disconnect(
+                        &connlib_client_shared::Error::Other("errors can't be serialized"),
+                    ),
+                    IpcServerMsg::OnUpdateResources(v) => callback_handler.on_update_resources(v),
+                    IpcServerMsg::OnSetInterfaceConfig { ipv4, ipv6, dns } => {
+                        callback_handler.on_set_interface_config(ipv4, ipv6, dns);
+                    }
+                }
+            }
+            Ok(())
+        });
+
+        let mut client = Self { task, tx };
+        let token = token.expose_secret().clone();
+        client
+            .send_msg(&IpcClientMsg::Connect {
+                api_url: api_url.to_string(),
+                token,
+            })
+            .await
+            .context("Couldn't send Connect message")?;
+        Ok(client)
+    }
+
     pub(crate) async fn reconnect(&mut self) -> Result<()> {
         self.send_msg(&IpcClientMsg::Reconnect)
             .await
